presets: Refactor preset list
Refactor the preset list into a new object that can manage multiple
properties, can add and remove items, and can be reordered.

Creates the UI actions for adding, removing and moving presets from the
preset list. Presets are no longer a static list, but can now be added
and removed as needed.

Fixes: #62, "Manually reordering presets in the list"

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>