onvif: Fix save and restore of presets
Preset save and restore was broken on ONVIF, partially because the
preset infrastructure was not able to store the preset 'token' used by
ONVIF instead of a simple preset number. Rework the ONVIF code to store
the token value in the preset structure instead of maintaining a
separate QMap lookup table.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>