Move the addition of command affects to stale settings from send_pending to receive
As mentioned in issue #170, adding a command's affects to stale_settings as the command is being sent causes the VISCA controller to generate an inquiry for the stale setting before the camera has had a chance to change the setting.  This was discovered particularly for Power On/Off setting.

By moving the code to the receive function after the camera sends a completion message for the command, we are assured that the result of the inquiry to refresh the stale setting will be the current setting.

Reported By: GeNTooGeek
Signed-off-by: GeNTooGeek <gentoogeek@3g5g.ca>