Add if statement to protect setting of wb_mode setting.
Currently implementation of set_settings doesn't proect the changing of  the wb_mode settings with an if statement to verify that the new settings include a new wb_mode.  The ended up generating extra requests for wb_mode when only requesting the change in power setting.

Added the if statement to protect it.

Reported by: GeNTooGeek
Signed-off-by: GeNTooGeek <gentoogeek@3g5g.ca>