Added Power On/Off Actions to PTZ Action Source
I wanted to be able to automatically turn my camera on and off with the use of Advanced Scene Switcher, but I noticed that the PTZ Action Source was missing Power On/Off as an action option.  Added the required code to add the items in the source properties, and other code to push the request to the device.

I followed the way the movement actions are implemented.  I later noticed that the PTZ Controls dialog includes a context menu option to turn on and off the camera's power.  While my implementation of the power actions is not done the same way as the way the context menu is implemented, the settings for the camera are properly updated after the PTZ Action is completed, so they play nicely together.

Reported by: GeNTooGeek
Signed-off-by: GeNTooGeek <gentoogeek@3g5g.ca>