Stop using deprecated obs_frontend_add_dock()
OBS Studio 30.0.0 deprecated the obs_frontend_add_dock() function and
advised plugins to use obs_frontend_add_dock_by_id() instead. However,
the new API takes the contents of the dock widget as an argument instead
of the dock widget itself, which requires rework of the widget
structure.

This commit removes the QDockWidget instance from the .ui file and
changes the PTZControls class to inheret from QWidget instead of
QDockWidget. It then reworks the instantiation to use the new API when
building against obs-studio 30.0.0 or higher. When building against
older versions it will create the needed QDockWidget before calling the
old API.

Fixes: #186
Signed-off-by: Grant Likely <grant.likely@secretlab.ca>