Update build scripts to latest obs-plugintemplate
This is a wholesale replacement of the build infrastructure with the
latest version from https://github.com/obsproject/obs-plugintemplate.
The new build baseline is now on commit 013ffcae, "cmake: Fix default
installation directory for Windows"

This commit disables UART camera control because QSerialPort is not
available in the prebuild OBS libraries, and it also disables joystick
support because SDL needs to be ported over into the new build
infrastructure. At least this commit gets the plugin building again and
working with a recent OBS Studio release.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>