clang-format: relax rules for this project
Some of the updated rules for clang-format cause big changes to the
imported source files, which makes is slightly harder to import updated
versions of external sources. Relax the rules for now. This can be
revisited if the scripts get updated to omit imported sources from
clang-format rules.

This patch also applies the small number of changes to match the updated
rules.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>