ui: Update imported obs property-view widget
Upstream OBS has made is much easier to integrate the property-view
widget. It is now an inclueable cmake subdirectory that builds cleanly,
which means this project can use a pristine copy with no local
modifications. It does however require moving the source to the shared/
directory so that the build scripts don't need to be modified.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>