Move qjoysticks into shared/
Like the OBS imported sources, move the qjoysticks source files out of
src/ and into shared/. Also add a local CMakeLists.txt so that can
contain the details of the QJoysticks build and keep it separate from
the main plugin.

Joystick support is only enable for Linux at the moment. More work is
required to enable on MacOS and Windows

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>