Move lock from toolbar to each device entry
Feedback from users is that the lock button on the toolbar isn't very
intuitive. It isn't clear that the lock is associated with the currently
selected camera, or what changes the lock state. Moving the lock button
into the list view beside the camera names is much clearer.

This change follows the pattern used in obs-studio source-tree.cpp, use
an delegate and overlay widget to put the lock button into the list.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>