ci: Load the vcvars64.bat environment when bulding plugin
On windows, building QtSerialPort (or any other Qt module) uses Ninja,
but ninja won't detect the VS tools unless the environment from
vcvars64.bat is loaded. Add a snippet of code to call vcvars64.bat, load
the settings into Powershell, and then run the build script.

It's not a pretty solution, and it will break when the path to Visual
Studio changes in the Windows runner. However, for now it gets the
plugin building for Windows.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>