ui: Move Auto select camera toggle into context menu
Since the Auto Select Camera feature is more of a setup configuration
and less of a live control, it does not need to be visible all the time
in the toolbar. Remove the Follow Pre and Follow Live buttons (which
have poor icons anyway) and add a toggleable Auto Select Camera icon in
the context menu.

This patch also drops the ability to follow the live camera when in
studio mode assuming it isn't used. However, it can easily be added back
if anyone needs it.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>