visco: Reduce timeout and drop warning
Messages are frequently lost in the visca protocol (particularly for the
UDP transport). The timeout timer is more of a retry timer that needs to
expire quickly so that the camera remains responsive. Drop the timeout
log message because that is scary and there isn't really a problem.

However, reduce the timeout period slighly to 20Hz. That will reduce the
traffic while still keeping the camera responsive.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>.