ui: Refresh instead of Reload when cam props change
The ReloadProperties() method clears and reloads all the widgets in the
properties view widget, which is a lot of churn, and resets the users
view, for changes that are usually a single integer or string change.
Instead, call the RefreshProperties() method via the event queue. This
both reduces the amount of work that needs to be done, and reduces the
chance of stalling the UI by running it through the event queue.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>