visca-udp: move datagram handling into main class
As part of reworking VISCA over IP protocol handling, move the datagram
parsing out of the ViscaUDPSocket helper class, and into the main
PTZViscaOverIP class. This is because parsing the datagram properly
requires some context from the main class to know what sequence numbers
to be looking for. The helper class, which is shared between multiple
instances, doesn't need that context.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>