visca: Correctly handle UDP sequence numbers
Rework the datagram processing to check the sequence numbers coming back
from the camera. If the sequence number doesn't match, then discard the
datagram. This protects against incorrectly interpreting the data in the
response datagram, for instance if the main visca code is expecting a
response from a different request.

Fixes: #170
Fixes: #159
Signed-off-by: Grant Likely <grant.likely@secretlab.ca>