settings: Move apply button into the designer form
There is no need to create the apply button manually. Move it into the
designer form so that all the UI elements are managed in the same place.
The one exception is the OBSPropertiesView widget, but that is because
it needs extra context when it is created that designer does not have.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>