joystick: make joystick buttons and axes configurable
Add joystick axis and button mapping to the settings dialog. The mapping
enables a small number of initial actions that can be mapped as the user
likes, including pan/tilt/zoom for the axes, and selecting previous/next
camera or preset, and recalling presets for the buttons. More actions
will get added in the future.

Fixes: #139
Signed-off-by: Grant Likely <grant.likely@secretlab.ca>