visca: Fix crash on unresolved hostname
The VISCA over UDP driver was unconditionally dereferencing an empty
list when the lookup_host callback was triggered. When a hostname was
resolved there was no problem, but on a lookup failure the addresses
list is empty, and the function segfaulted. Fix by checking to see if
the list isn't empty before dereferencing it.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>