settings: Change layout of joystick controls
This commit reworks the layout of the general settings tab. It puts the
entire configuration into a scroll area and move the general settings to
the top, above the joystick settings. Finally, it puts the joystick axis
and button mappings into a multi column grid so that it doesn't take up
as much space. With this change it is easier to see all the joystick
mappings at once, and it is easier to find the non-joystick related
settings.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>