Partial Revert "Put properties action into the camera context menu"
This partially reverts commit 98d25fd7f1a6e9d28bf916860fe17c8077d72849.
The properties button in the toolbar was missed by users and is used
frequently enough to warrent being on the toolbar. Also, from watching
user videos, not everyone realizes that there is a context menu
available in the cameras list, or that it can be opened by double
clicking on the camera. Bringing back the properties button makes
getting into the configuration dialog more intuitive.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>