visca/udp: check if host field is an IP address
If the host field is an IP address, then the call to lookupHost() will
delay for some time while it tries to do a reverse hostname lookup. This
isn't what is wanted, and lookupHost() doesn't need to be called to
parse an IP address field. Fix this by trying first to parse the host
field as an IP address, and only call lookupHost() if that fails (which
it will if the host field contains a hostname.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>