visca/udp: Handle cameras which reply on a slot without an ack
Some cameras will answer an inquiry using a slot number without sending
an ack packet first. The code was depending on the ack packet to work
out which slot the reply would arrive on and ignore the reply if the ack
was not received first. Simplify the code to not depend on the ack, and
accept any message who's sequence number matches either the specified
slot or slot 0.

The changes affect both core VISCA code and the VISCA-over-IP code as
both were looking for the ack packet explicitly. An advantage of this
chagne is that it simplifies the VISCA-over-IP to parse as little as
possible of the VISCA packet in the UDP code and instead let the core
VISCA logic handle it as it does for serial and TCP.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>