visca: add scanning of all commands
Adds a log of the latest response from each inquiry sent which is
helpful for debugging the camera protocol. This commit adds two buttons
to the camera configuration dialog. One to initiate a scan of all the
possible inquiry commands, and another to dump the last reply buffer to
the log.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>