Add device option to invert pan/tilt/zoom/focus axis
Sometimes the default visca directions don't work for a camera. This
could be because the camera was installed inverted, the video stream
is flipped, or the camera simply doesn't follow direction conventions.
This commit adds options to invert the direction of the pan, tilt, zoom
and focus controls so that the control panel uses consistent directions
for all devices.

Fixes: #11
Signed-off-by: Grant Likely <grant.likely@secretlab.ca>