joystick: Map hotkeys to buttons
Rework the joystick button code to directly map to OBS hotkeys. This
makes it possible to map any OBS hotkey to a joystick button, and not be
restricted only to what the plugin is able to implement.

With this commit only the frontend hotkeys are made accessible. Some
more work is required to map hotkeys from other hotkey registerers.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>