Add hotkey combination for selecting previous/next scene
This isn't really an PTZ feature, but OBS-Studio doesn't already have
hotkeys for navigating the scene list and it is important to have for
complete joystick control. This adds two new hotkeys for navigating the
scene list; based on the display order, the 'ScenePrev' and 'SceneNext'
hotkeys select the previous and next scene respectively.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>