ui: Fix layout of device list
The device list QListView wasn't allocating enough space for the lock
button on OBS Studio 31.1.0 beta due to the way sizeHint was getting
calculated. It was using incorrect code copied from obs-studio that was
fixed in commit dbee6d27, "frontend: Adjust styling for Sources list".
This commit copies the corrected code over and adds a calculation to
keep every row consistently sized for the lock button.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>