ui: Replace QPushButton with QToolButton
QPushButton makes assumptions about how it is used, and adds padding and
other size constraints for that usage. The direction controls however
need to be layed out in a neat grid. QToolButton is a better choice as
it provides a clickable button, but it doesn't impose any extra layout
constraints.

Fixes: #245
Signed-off-by: Grant Likely <grant.likely@secretlab.ca>