ui: Remove left and right margins from buttons
In the 31.0.x OBS Studio series, the theme sets top and bottom margins
on buttons to 0, but leaves the padding in place left and right. When
laying out the buttons in a grid this makes the left and right gutters
visibly larger than the top and bottom ones. Fix the problem by manually
specifying a stylesheet for each grid button.

This problem doesn't exist in the latest 31.1.0 beta release. This is a
temporary fix until support is dropped for 31.0.x.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>