cmake: Get version from git tag
The canonical source of version information is the git tag instead of
buildspec. The patch imports the git version decoding scripts from OBS Studio
and adapts them for the plugin.

The `cmake/common/versionconfig.cmake` file in particular is copied over from
obs-studio with a couple of light modifications. First, all the {OBS_,_obs_}*
variables are renamed to {PACKAGE,_package_}* because this is not the OBS
studio package, and the tag parsing regex expressions are modified to handle a
`v` prefix on the tag as is used in this project.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>